Refine cip image structure creating function and manipulating buffer function. Align all function name or variable name to bottom line format. Also add new section to convert image color space.
In this commit, we align all function name and declare they with bottom
line format. We also refine functions which create basic cip image data,
like (create, copy, delete), and it's supported to using this
implemented function to create cip image data or free it. Finally, we
add convert cip image color space in this commit, but it's still not
completed yet, and those function still need to be integrate with BMP
file storing function.